- Create new file containing methods to remove list items - Use array filter method to remove completed items from list - Add event listener to checkbox and check completed items - Update local storage
